チュートリアル完了

+++ laravel9-build-template/src/LaravelTestProject/app/Http/Controllers/FolderController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Folder;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Log;
+use App\Models\Folder;
 use App\Http\Requests\CreateFolder;
 use App\Http\Requests\EditFolder;
-use Illuminate\Support\Facades\Auth;
 
 class FolderController extends Controller
 {
@@ -18,28 +20,42 @@ class FolderController extends Controller
      */
     public function showCreateForm()
     {
-        $folders = Auth::user()->folders;
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $user->folders;
 
-        return view('folders.create', compact('folders'));
+            return view('folders/create');
+
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showCreateForm: ' . $e->getMessage());
+        }
     }
 
-        /**
+    /**
      *  【フォルダの作成機能】
      *
      *  POST /folders/create
-     *  @param Request $request （リクエストクラスの$request）
+     *  @param CreateFolder $request （Requestクラスの機能は引き継がれる）
      *  @return \Illuminate\Http\RedirectResponse
+     *  @var App\Http\Requests\CreateFolder
      */
     public function create(CreateFolder $request)
     {
-        $folder = new Folder();
-        $folder->title = $request->title;
+        try {
+            $folder = new Folder();
+            $folder->title = $request->title;
 
-         /** @var App\Models\User **/
-        $user = Auth::user();
-        $user->folders()->save($folder);
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $user->folders()->save($folder);
 
-        return redirect()->route('tasks.index', ['folder' => $folder->id]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id,
+            ]);
+        } catch (\Exception $e) {
+            Log::error('Error FolderController in create: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -51,34 +67,43 @@ class FolderController extends Controller
      */
     public function showEditForm(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        return view('folders.edit', [
-            'folder_id' => $folder->id,
-            'folder_title' => $folder->title,
-        ]);
+            return view('folders/edit', [
+                'folder_id' => $folder->id,
+                'folder_title' => $folder->title,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in showEditForm: ' . $e->getMessage());
+        }
     }
 
     /**
-     * 【フォルダの編集機能】
+     *  【フォルダの編集機能】
+     *
      *  POST /folders/{folder}/edit
      *  @param Folder $folder
-     *  @param CreateFolder $request
-     * @return \Illuminate\Http\RedirectResponse
+     *  @param EditTask $request
+     *  @return \Illuminate\Http\RedirectResponse
      */
-    public function edit(Folder $folder, EditFolder $request) {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+    public function edit(Folder $folder, EditFolder $request)
+    {
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
+            $folder->title = $request->title;
+            $folder->save();
 
-        $folder->title = $request->title;
-        $folder->save();
-
-        return redirect()->route('tasks.index', [
-            'folder' => $folder->id
-        ]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in edit: ' . $e->getMessage());
+        }
     }
 
     /**
@@ -90,19 +115,22 @@ class FolderController extends Controller
      */
     public function showDeleteForm(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        return view('folders.delete', [
-            'folder_id' => $folder->id,
-            'folder_title' => $folder->title,
-        ]);
+            return view('folders/delete', [
+                'folder_id' => $folder->id,
+                'folder_title' => $folder->title,
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error in showDeleteForm: ' . $e->getMessage());
+        }
     }
 
     /**
      *  【フォルダの削除機能】
-     *  機能：フォルダが削除されたらDBから削除し、フォルダ一覧にリダイレクトする
      *
      *  POST /folders/{folder}/delete
      *  @param Folder $folder
@@ -110,17 +138,25 @@ class FolderController extends Controller
      */
     public function delete(Folder $folder)
     {
-        /** @var App\Models\User **/
-        $user = Auth::user();
-        $folder = $user->folders()->findOrFail($folder->id);
+        try {
+            /** @var App\Models\User **/
+            $user = Auth::user();
+            $folder = $user->folders()->findOrFail($folder->id);
 
-        $folder->tasks()->delete();
-        $folder->delete();
+            $folder = DB::transaction(function () use ($folder) {
+                if($folder) throw new \Exception('500');
+                $folder->tasks()->delete();
+                $folder->delete();
+                return $folder;
+            });
 
-        $folder = Folder::first();
+            $folder = Folder::first();
 
-        return redirect()->route('tasks.index', [
-            'folder' => $folder->id
-        ]);
+            return redirect()->route('tasks.index', [
+                'folder' => $folder->id
+            ]);
+        } catch (\Throwable $e) {
+            Log::error('Error FolderController in delete: ' . $e->getMessage());
+        }
     }
 }
